removed tailing slashes

+++ run-on-k8.sh
@@ -1,11 +1,28 @@
 
 # these files need to be in the spark classpath in order for spark-submit to be able to upload/download the fat jar from s3
-wget https://repo1.maven.org/maven2/org/bouncycastle/bcprov-jdk15on/1.70/bcprov-jdk15on-1.70.jar && mv bcprov-jdk15on-1.70.jar /opt/spark/jars/
-wget https://repo1.maven.org/maven2/org/bouncycastle/bcpkix-jdk15on/1.70/bcpkix-jdk15on-1.70.jar && mv bcpkix-jdk15on-1.70.jar /opt/spark/jars/
-wget https://repo1.maven.org/maven2/org/apache/hadoop/hadoop-aws/3.3.4/hadoop-aws-3.3.4.jar && mv hadoop-aws-3.3.4.jar /opt/spark/jars/
-wget https://repo1.maven.org/maven2/com/amazonaws/aws-java-sdk-bundle/1.11.901/aws-java-sdk-bundle-1.11.901.jar && mv aws-java-sdk-bundle-1.11.901.jar /opt/spark/jars/
 
-$HOST_MOUNT=/home/warren/storage-mount/amicusearch-v2-data/
+if [ ! -f "/opt/spark/jars/bcprov-jdk15on-1.70.jar" ]; then
+  wget https://repo1.maven.org/maven2/org/bouncycastle/bcprov-jdk15on/1.70/bcprov-jdk15on-1.70.jar && mv bcprov-jdk15on-1.70.jar /opt/spark/jars/
+else
+  echo "bcprov-jdk15on-1.70.jar already exists"
+fi
+if [ ! -f "/opt/spark/jars/bcpkix-jdk15on-1.70.jar" ]; then
+  wget https://repo1.maven.org/maven2/org/bouncycastle/bcpkix-jdk15on/1.70/bcpkix-jdk15on-1.70.jar && mv bcpkix-jdk15on-1.70.jar /opt/spark/jars/
+else
+  echo "bcpkix-jdk15on-1.70.jar already exists"
+fi
+if [ ! -f "/opt/spark/jars/hadoop-aws-3.3.4.jar" ]; then
+  wget https://repo1.maven.org/maven2/org/apache/hadoop/hadoop-aws/3.3.4/hadoop-aws-3.3.4.jar && mv hadoop-aws-3.3.4.jar /opt/spark/jars/
+else
+  echo "hadoop-aws-3.3.4.jar already exists"
+fi
+if [ ! -f "/opt/spark/jars/aws-java-sdk-bundle-1.11.901.jar" ]; then
+  wget https://repo1.maven.org/maven2/com/amazonaws/aws-java-sdk-bundle/1.11.901/aws-java-sdk-bundle-1.11.901.jar && mv aws-java-sdk-bundle-1.11.901.jar /opt/spark/jars/
+else
+  echo "aws-java-sdk-bundle-1.11.901.jar already exists"
+fi
+
+HOST_MOUNT=/home/warren/storage-mount/amicusearch-v2-data/
 
 nohup spark-submit \
   --master k8s://https://127.0.0.1:6443 \
@@ -27,4 +44,4 @@ nohup spark-submit \
   --conf spark.hadoop.fs.s3a.access.key=$AWS_ACCESS_KEY_ID \
   --conf spark.hadoop.fs.s3a.secret.key=$AWS_SECRET_ACCESS_KEY \
   s3a://amicusearch/etl/AmicusearchETL.jar \
-  --mode courtlistener --env local --states FL,NY --includeFederal true > spark.log 2>&1 &
+  --mode courtlistener --env dev --states FL,NY --includeFederal true > spark.log 2>&1 &
